Solid pricniple changes
