+++ java/src/main/java/com/codurance/training/tasks/model/Task.java
@@ -26,4 +26,13 @@ public final class Task {
     public void setDone(boolean done) {
         this.done = done;
     }
+
+    @Override
+    public String toString() {
+        return "Task{" +
+                "id=" + id +
+                ", description='" + description + '\'' +
+                ", done=" + done +
+                '}';
+    }
 }
